+++ client/testing.go
@@ -9,8 +9,8 @@ import (
 	"github.com/hashicorp/nomad/client/fingerprint"
 	"github.com/hashicorp/nomad/command/agent/consul"
 	"github.com/hashicorp/nomad/helper/testlog"
-	"github.com/hashicorp/nomad/pluginutils/catalog"
-	"github.com/hashicorp/nomad/pluginutils/singleton"
+	"github.com/hashicorp/nomad/helper/pluginutils/catalog"
+	"github.com/hashicorp/nomad/helper/pluginutils/singleton"
 	testing "github.com/mitchellh/go-testing-interface"
 )
