+++ command/agent/plugins.go
@@ -3,9 +3,9 @@ package agent
 import (
 	"fmt"
 
-	"github.com/hashicorp/nomad/pluginutils/catalog"
-	"github.com/hashicorp/nomad/pluginutils/loader"
-	"github.com/hashicorp/nomad/pluginutils/singleton"
+	"github.com/hashicorp/nomad/helper/pluginutils/catalog"
+	"github.com/hashicorp/nomad/helper/pluginutils/loader"
+	"github.com/hashicorp/nomad/helper/pluginutils/singleton"
 )
 
 // setupPlugins is used to setup the plugin loaders.
