+++ helper/pluginutils/catalog/register.go
