+++ helper/pluginutils/catalog/register_testing.go
