+++ helper/pluginutils/hclspecutils/spec_funcs.go
