+++ helper/pluginutils/hclutils/util.go
