+++ helper/pluginutils/loader/filter_windows.go
