+++ helper/pluginutils/loader/loader_test.go
