+++ helper/pluginutils/loader/plugin_test.go
