+++ helper/pluginutils/loader/testing.go
