+++ plugins/base/client.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/hashicorp/nomad/plugins/base/proto"
 	"github.com/hashicorp/nomad/plugins/shared/hclspec"
-	"github.com/hashicorp/nomad/pluginutils/grpcutils"
+	"github.com/hashicorp/nomad/helper/pluginutils/grpcutils"
 )
 
 // BasePluginClient implements the client side of a remote base plugin, using
