+++ plugins/drivers/client.go
@@ -16,7 +16,7 @@ import (
 	"github.com/hashicorp/nomad/plugins/shared/hclspec"
 	pstructs "github.com/hashicorp/nomad/plugins/shared/structs"
 	sproto "github.com/hashicorp/nomad/plugins/shared/structs/proto"
-	"github.com/hashicorp/nomad/pluginutils/grpcutils"
+	"github.com/hashicorp/nomad/helper/pluginutils/grpcutils"
 	"google.golang.org/grpc/status"
 )
